# (#6059) - use batch cursor in allDocs()