I rebased this on #6056 to avoid bitrotting myself.

This extends our use of `getAll()`/`getAllKeys()` to `allDocs()`, in addition to the work already done to use it in `changes()`. ~~This was a bit trickier because the two have slightly different use cases, but this is documented in the code comments for `runBatchedCursor.js`.~~

~~There is an additional optimization we can do here where we avoid `getAllKeys()` entirely for `allDocs()` (it's not necessary, since unlike the `by-seq-store` for the `doc-store` we can derive the key from the value; it's not an autoincrementing key). However I'll save that for a future PR.~~

_**edit:** I have a new implementation using batched cursors or a regular `getAll` depending on the scenario._

Benchmarks are TBD but in principle this is faster with no increased memory costs since we're still keeping the batch size limited based on `limit`.