@nolanlawson 2016-12-24
Interestingly this is a perf boost relative to master in Firefox, but a regression in Chrome (chrome 119ms -> 868ms, firefox 496ms -> 157ms for an `allDocs()` test I'm running). However, that's counting the removal of the `docCount` optimization as well.