@nolanlawson 2016-12-24
Hm, based on these recent results I'm starting to suspect that `objectStore.count()` is just expensive in Chrome whereas it's cheap in Firefox (something to do with LevelDB vs SQLite?). In any case we may want to bring the docCount optimization back, but have it inside the store instead of in-memory so that it works across threads.

I am still happy with this PR but I'm sad that the docCount optimization helped so much in Chrome.