@nolanlawson 2016-12-24
OK, in order to compare apples to apples, I compared using the new all-docs-startkey-endkey using _just_ the commits in this PR (and using #6056 as the baseline, bumped up to 50 iterations in order to be more sure):

- Chrome: 6740ms -> 7583ms
- Firefox: 2933ms -> 1092ms

So oddly it is a perf regression in Chrome. Apparently it's faster to fetch 10 documents using a cursor than to fetch them all at once using a single `getAll()`.

This is a weird result, but I still think this is a good PR. We can improve Chrome perf by doing #6068. Also even if it's a bit slower, I can see from Dev Tools timelines that it spends less time on the UI thread because it has fewer callbacks. So a different benchmark may be fairer.