@nolanlawson 2016-12-24
Small bug in `descending=true` mode that was only being tested in the mapreduce tests... so I added integration tests for it and fixed it. Should be green now.