@nolanlawson 2016-12-24
Collapsed commits so it's easier to read.