@nolanlawson 2016-12-27
Rebased, will wait for a green before merging.